Conflicts:
	db/db_impl.cc
	db/db_test.cc
	db/internal_stats.cc
	db/internal_stats.h
	db/version_edit.cc
	db/version_edit.h
	db/version_set.cc
	include/rocksdb/options.h
	util/options.cc

+++ db/corruption_test.cc
@@ -387,7 +387,7 @@ TEST(CorruptionTest, FileSystemStateCorrupted) {
     DBImpl* dbi = reinterpret_cast<DBImpl*>(db_);
     std::vector<LiveFileMetaData> metadata;
     dbi->GetLiveFilesMetaData(&metadata);
-    ASSERT_GT(metadata.size(), 0);
+    ASSERT_GT(metadata.size(), size_t(0));
     std::string filename = dbname_ + metadata[0].name;
 
     delete db_;
